08 Donnerstag/mail/src/main/java/com/example/mail/MailService.java
package com.example.mail;

import org.springframework.mail.SimpleMailMessage;
import org.springframework.mail.javamail.JavaMailSender;
import org.springframework.stereotype.Component;

@Component
public class MailService {

    private JavaMailSender mailSender;

    public MailService(JavaMailSender mailSender) {
        this.mailSender = mailSender;
    }
    public void sendMail() {
        SimpleMailMessage mailMessage = new SimpleMailMessage();
        mailMessage.setFrom("sender@test.org");
        mailMessage.setTo("receiver@test.org");
        mailMessage.setSubject("Subject");
        mailMessage.setText("Text");
        mailSender.send(mailMessage);
    }
}
